Changes to color pattern in footer
edits to line 1512

change div color to match mockup color pattern and coordinate with the rest of the pages. Since footer == dark blue
I changed the section directly above the footer to be a dark grey instead. This way the footer color remains consistent throughout
all of the available pages. Personally i like the dark color more, however if you're not a fan of this change just go to
line 1512 to change the section above the footer and line 1536 to change the footer color respectively.